Backend

+++ apps/api/src/models/product.py
@@ -1,5 +1,5 @@
 import uuid
-from sqlalchemy import Column, String, Numeric, Boolean, DateTime
+from sqlalchemy import Column, String, Float, Boolean, DateTime
 from sqlalchemy.sql import func
 from src.core.database import Base
 
@@ -13,8 +13,8 @@ class Product(Base):
     name = Column(String, nullable=False)
     sku = Column(String, unique=True, index=True, nullable=False)
     category = Column(String, default="hardware", nullable=True) # e.g. "hardware", "service", "software"
-    sales_price = Column(Numeric(10, 2), nullable=False)
-    cost = Column(Numeric(10, 2), nullable=False)
+    sales_price = Column(Float, nullable=False)
+    cost = Column(Float, nullable=False)
     active = Column(Boolean, default=True)
     
     created_at = Column(DateTime(timezone=True), server_default=func.now())
